feat(amplify): update dependencies and refactor function structure
This commit introduces the axios library as a dependency in the package.json files for both the main amplify project and the cancelPlan, planManagement, and webHookPlan functions. It also refactors the entry points for these functions to simplify their structure by removing the 'src/' prefix from the handler file paths. Additionally, it deletes unused handler files and their corresponding package-lock.json files, streamlining the codebase and improving maintainability.
