+++ amplify/functions/cancelPlan/resource.ts
@@ -2,7 +2,7 @@ import { defineFunction, secret } from '@aws-amplify/backend'
 
 export const cancelPlan = defineFunction({
   name: 'cancelPlan',
-  entry: 'src/handler.ts',
+  entry: 'handler.ts',
   environment: {
     MERCADOPAGO_ACCESS_TOKEN: secret('MERCADOPAGO_ACCESS_TOKEN'),
   },
