--- amplify/functions/cancelPlan/src/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "src",
-  "version": "1.0.0",
-  "main": "handler.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "axios": "^1.7.9"
-  }
-}
