+++ amplify/functions/webHookPlan/handler.ts
@@ -10,7 +10,7 @@ import { Amplify } from 'aws-amplify'
 import { generateClient } from 'aws-amplify/data'
 import { getAmplifyDataClientConfig } from '@aws-amplify/backend/function/runtime'
 import { env } from '$amplify/env/hookPlan'
-import { type Schema } from '../../../data/resource'
+import { type Schema } from '../../data/resource'
 
 const { resourceConfig, libraryOptions } = await getAmplifyDataClientConfig(env)
 Amplify.configure(resourceConfig, libraryOptions)
