+++ amplify/package.json
@@ -1,3 +1,6 @@
 {
-  "type": "module"
+  "type": "module",
+  "dependencies": {
+    "axios": "^1.9.0"
+  }
 }
